backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
networkx==3.2
automata-lib==8.2.0
jinja2==3.1.2
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1 